requirements.txt
--index-url https://phonepe.mycloudrepo.io/public/repositories/phonepe-pg-sdk-python
--extra-index-url https://pypi.org/simple

APScheduler==3.10.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.28.80
botocore==1.31.80
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
crypto==1.4.1
dataclasses-json==0.5.8
dj-database-url==2.1.0
Django==4.1.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.10.0
django-rest-framework==0.1.0
django-storages==1.13.2
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
environs==9.5.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
Naked==0.1.32
packaging==23.1
paytmchecksum==1.7.0
phonepe-sdk==1.1.0
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
psycopg2==2.9.9
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0
requests==2.31.0
resend==0.6.0
rest-framework-simplejwt==0.0.2
rsa==4.9
s3transfer==0.7.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.12
